Use release_version in recreate benchmark message
The correlation key was configured to the string literal
"release_version", but it should've used the actual value of the
release_version variable.

I've also expanded the user task with a form to explain what to do and
added a note to the message to see the name and correlation key.