Improve update long running benchmarks form
The text was inconsistently using its own naming scheme. In the
meantime, a new naming scheme was adopted:
- `release-1-x` for a major version, that is kept up-to-date with minor
  versions
- `release-1-1-x` for a minor version, that is kept up-to-date with
  patch versions. Since we support 2 of these, 2 should be running.
- `release-1-1-alpha1` for an alpha version.

In addition, this replaces some manual changes by using `sed` as an
easier way to make those changes.

Lastly, this adds the task of updating the major `release-1-x` version
to newer minor versions.