fix: add upgrade zeebe version description for patches on previous releases
For patches on previous releases dependabot will not create a PR to the stable branch. This is a manual step which was not documented in the release process.
I've also added a not that these kind of releases are not necessary for version lower than 1.4.0. This is because 1.4.0 will be the first official stable release. No stable branches exist for earlier versions