fix(post_release): rm long running major release
The long running major release served no purpose. It was often not kept
up-to-date with the latest patch, and if it was then there was already
another long running cluster through the latest minor release benchmark.

This commit removes all mentions of this benchmark from the update
benchmarks form. It also tries to clean up parts of the text.