fix: stablize flaky test case
The QA Process Test was a bit flaky. This could happen because the
engine has a bit of idle state just after creating the timer, which
was expected to be triggered immediately. However, it doesn't always
because of racing.

I think that actually waiting for an idle state, then increasing the
time to allow triggering the timer, waiting for another idle state
(possibly the task, possibly the timer event that hasn't yet triggered)
is enough. Note that it would be better to have a way to await the
triggered timer event, but until we have that I think this will work
(mostly).